Removed integrations folder
